pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "aster-multimedia"
version = "0.1.0"
description = "Modular toolkit for running multimodal generative AI workflows locally"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Local AI Toolkit"}
]
license = {text = "MIT"}
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "pyyaml>=6.0",
  "click>=8.1"
]

[project.optional-dependencies]
text = [
  "torch",
  "transformers",
  "accelerate"
]
image = [
  "torch",
  "diffusers",
  "transformers",
  "accelerate",
  "safetensors",
  "opencv-python",
  "pillow"
]
video = [
  "torch",
  "diffusers",
  "transformers",
  "accelerate",
  "imageio[ffmpeg]",
  "opencv-python"
]

[project.scripts]
aster = "aster_multimedia.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
